+++ Modules/Microsoft365DSC/Modules/M365DSCExportUtil.psm1
@@ -20,6 +20,15 @@ foreach ($template in $jsonContent.templates.psobject.Properties)
 {
     $Script:RelationTemplates.templates[$template.Name] = $template.Value
 }
+$allResourcesArgumentCompleter = Get-M365DSCAllResources
+Register-ArgumentCompleter -CommandName Export-M365DSCConfiguration -ParameterName Components -ScriptBlock {
+    param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameters)
+    $resources = $allResourcesArgumentCompleter -like "$wordToComplete*"
+    foreach ($resource in $resources)
+    {
+        [System.Management.Automation.CompletionResult]::new($resource, $resource, 'ParameterValue', $resource)
+    }
+}
 
 <#
 .SYNOPSIS